test: User관련 테스트 코드 수정
