--- src/test/java/techcourse/myblog/domain/vo/user/UserBasicInfoTests.java
@@ -1,17 +1,0 @@
-package techcourse.myblog.domain.vo.user;
-
-import org.junit.jupiter.api.Test;
-import techcourse.myblog.user.User;
-import techcourse.myblog.user.UserBasicInfo;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class UserBasicInfoTests {
-	@Test
-	void valueOfUser() {
-		UserBasicInfo userBasicInfo = new UserBasicInfo("tiber", "tiber@naver.com");
-		User user = userBasicInfo.valueOfUser();
-		assertThat(user.getUsername()).isEqualTo(userBasicInfo.getUsername());
-		assertThat(user.getEmail()).isEqualTo(userBasicInfo.getEmail());
-	}
-}
